+++ mmseg/models/decode_heads/__init__.py
@@ -2,7 +2,6 @@
 from .ann_head import ANNHead
 from .apc_head import APCHead
 from .aspp_head import ASPPHead
-from .cat_head import CATSegHead
 from .cc_head import CCHead
 from .da_head import DAHead
 from .ddr_head import DDRHead
@@ -43,5 +42,5 @@ __all__ = [
     'SETRMLAHead', 'DPTHead', 'SETRMLAHead', 'SegmenterMaskTransformerHead',
     'SegformerHead', 'ISAHead', 'STDCHead', 'IterativeDecodeHead',
     'KernelUpdateHead', 'KernelUpdator', 'MaskFormerHead', 'Mask2FormerHead',
-    'LightHamHead', 'PIDHead', 'DDRHead', 'CATSegHead'
+    'LightHamHead', 'PIDHead', 'DDRHead'
 ]
