+++ mmseg/models/necks/__init__.py
@@ -1,6 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .cat_aggregator import (AggregatorLayer, CATSegAggregator,
-                             ClassAggregateLayer, SpatialAggregateLayer)
 from .featurepyramid import Feature2Pyramid
 from .fpn import FPN
 from .ic_neck import ICNeck
@@ -9,7 +7,5 @@ from .mla_neck import MLANeck
 from .multilevel_neck import MultiLevelNeck
 
 __all__ = [
-    'FPN', 'MultiLevelNeck', 'MLANeck', 'ICNeck', 'JPU', 'Feature2Pyramid',
-    'CATSegAggregator', 'SpatialAggregateLayer', 'ClassAggregateLayer',
-    'AggregatorLayer'
+    'FPN', 'MultiLevelNeck', 'MLANeck', 'ICNeck', 'JPU', 'Feature2Pyramid'
 ]
