+++ mmseg/models/utils/__init__.py
@@ -7,8 +7,7 @@ from .make_divisible import make_divisible
 from .ppm import DAPPM, PAPPM
 from .res_layer import ResLayer
 from .se_layer import SELayer
-from .self_attention_block import (FullAttention, LinearAttention,
-                                   SelfAttentionBlock)
+from .self_attention_block import SelfAttentionBlock
 from .shape_convert import (nchw2nlc2nchw, nchw_to_nlc, nlc2nchw2nlc,
                             nlc_to_nchw)
 from .up_conv_block import UpConvBlock
