+++ mmseg/utils/__init__.py
@@ -9,8 +9,6 @@ from .class_names import (ade_classes, ade_palette, bdd100k_classes,
                           potsdam_palette, stare_classes, stare_palette,
                           synapse_classes, synapse_palette, vaihingen_classes,
                           vaihingen_palette, voc_classes, voc_palette)
-from .clip_templates import (IMAGENET_TEMPLATES, IMAGENET_TEMPLATES_SELECT,
-                             IMAGENET_TEMPLATES_SELECT_CLIP, ViLD_templates)
 from .collect_env import collect_env
 from .io import datafrombytes
 from .misc import add_prefix, stack_batch
@@ -30,6 +28,5 @@ __all__ = [
     'loveda_palette', 'potsdam_palette', 'vaihingen_palette', 'isaid_palette',
     'stare_palette', 'dataset_aliases', 'get_classes', 'get_palette',
     'datafrombytes', 'synapse_palette', 'synapse_classes', 'bdd100k_classes',
-    'bdd100k_palette', 'IMAGENET_TEMPLATES', 'IMAGENET_TEMPLATES_SELECT',
-    'IMAGENET_TEMPLATES_SELECT_CLIP', 'ViLD_templates'
+    'bdd100k_palette'
 ]
