+++ projects/CAT-Seg/configs/_base_/datasets/ade20k_384x384.py
