+++ projects/CAT-Seg/configs/_base_/datasets/coco-stuff164k_384x384.py
