+++ projects/CAT-Seg/configs/_base_/datasets/pascal_context_59_384x384.py
