+++ projects/CAT-Seg/mmseg/utils/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .self_attention_block import FullAttention, LinearAttention
+
+__all__ = ['FullAttention', 'LinearAttention']
