+++ projects/CAT-Seg/mmseg/utils/bpe_vocab/bpe_simple_vocab_16e6.txt.gz
