+++ projects/CAT-Seg/mmseg/utils/clip_model.py
