+++ projects/CAT-Seg/mmseg/utils/clip_wrapper.py
