+++ projects/CAT-Seg/utils/clip_templates.py
